Move [DocumentationVersion] extension into DocumentationVersion file
